Rearrange components and folder structure.
