+++ src/components/MaplibreMap.jsx
@@ -13,7 +13,7 @@ import useResponsive from '../hooks/useResponsive'
 import customIcon from '../assets/map-pin.png'
 import { MAIN_MAP_ID } from '../constants/constants'
 
-import MapAndTableControls from './MapAndTableControls/MapAndTableControls'
+import ViewAndZoomControls from './ViewAndZoomControls/ViewAndZoomControls'
 
 const defaultLon = -79
 const defaultLat = 32
@@ -250,8 +250,8 @@ const MaplibreMap = ({ mapRef, view, setView }) => {
     }
   }
 
-  const mapAndTableControlsWrapper = mapRef.current ? (
-    <MapAndTableControls map={mapRef.current.getMap()} view={view} setView={setView} />
+  const viewAndZoomControlsWrapper = mapRef.current ? (
+    <ViewAndZoomControls map={mapRef.current.getMap()} view={view} setView={setView} />
   ) : null
 
   const hideMapStyleLayers = (map) => {
@@ -309,7 +309,7 @@ const MaplibreMap = ({ mapRef, view, setView }) => {
         interactiveLayerIds={['sites-cluster-layer', 'sites-unclustered-layer']}
         attributionControl={false}
       >
-        {mapAndTableControlsWrapper}
+        {viewAndZoomControlsWrapper}
         {isDesktopWidth && (
           <AttributionControl
             compact={true}
